pyproject.toml
[project]
name = "langchain-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langchain-groq",
    "python-dotenv",
    "pydantic",
    ]
